webapp_staging/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.43
alembic==1.16.5
asyncpg==0.29.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
python-multipart==0.0.9
requests==2.32.3
pydantic==2.9.2
typing-extensions==4.12.2
greenlet==3.2.4
MarkupSafe==3.0.2
Mako==1.3.10
